operator/status: clear azure path fix job conditions on operator removal
there is sometimes a race condition when setting .spec.managementState
back and forth between Removed and Managed where the azure path fix
controller will kick off the job, but the resources needed for the job
to run will get removed (as expected) before the job can finish and its
controller update the operator's progressing condition to reflect that.

this has been happening often in the TestLeaderElection e2e test. this
test does not wait for the image registry operator to become available
before removing it, and this is what triggers the race. a customer has
reported a similar issue, although their error was slightly different
and proved harder to reproduce. this commit should fix both problems.

clearing the status is most important when users upgrade from a version
of the azure path fix controller that deploys the job. OCP versions that
do not deploy the job should not have the problem, and including this
code in them should be harmless.
in OCP >= 4.17 this job is no longer deployed, though the controller is
kept it can probably be safely removed in OCP >= 4.18.